requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beautifulsoup4==4.14.3
bleach==6.3.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.2
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
Markdown==3.8
MarkupSafe==3.0.2
openai==1.78.1
passlib==1.7.4
pillow==12.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
resend==2.10.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
fastapi-csrf-protect==1.0.7
itsdangerous==2.2.0
pydantic-settings==2.12.0